requirements.txt
torch
torchvision
pcl
open3d
plotly
pyyaml
tqdm
wandb>=0.12.10